requirements.txt
python-telegram-bot[socks]==20.a1
requests
